+++ client/src/components/Block/index.tsx
@@ -1,7 +1,7 @@
 import { BLOCK_EVENT } from '@wabinar/constants/socket-message';
 import ee from 'components/Mom/EventEmitter';
 import { memo, useEffect, useRef, useState } from 'react';
-import useSocketContext from 'src/hooks/useSocketContext';
+import useSocketContext from 'src/hooks/context/useSocketContext';
 
 import QuestionBlock from './QuestionBlock';
 import TextBlock from './TextBlock';
