+++ client/src/components/common/Selector/Dropdown/Item.tsx
@@ -1,4 +1,4 @@
-import { useDropdownContext } from 'src/hooks/useDropdownContext';
+import useDropdownContext from 'src/hooks/context/useDropdownContext';
 
 interface ItemProps {
   id: number;
