+++ client/src/hooks/context/useMeetingContext.ts
@@ -1,9 +1,9 @@
 import { useContext } from 'react';
 import ERROR_MESSAGE from 'src/constants/error-message';
-import { SelectedMomContext } from 'src/contexts/selected-mom';
+import MeetingContext from 'src/contexts/meeting';
 
-export default function useSelectedMom() {
-  const context = useContext(SelectedMomContext);
+export default function useMeetingContext() {
+  const context = useContext(MeetingContext);
 
   if (!context) throw new Error(ERROR_MESSAGE.OUT_OF_CONTEXT_SCOPE);
