+++ client/src/hooks/context/useUserContext.ts
@@ -1,9 +1,9 @@
 import { useContext } from 'react';
 import ERROR_MESSAGE from 'src/constants/error-message';
-import MeetingContext from 'src/contexts/meeting';
+import UserContext from 'src/contexts/user';
 
-export function useMeetingContext() {
-  const context = useContext(MeetingContext);
+export default function useUserContext() {
+  const context = useContext(UserContext);
 
   if (!context) throw new Error(ERROR_MESSAGE.OUT_OF_CONTEXT_SCOPE);
